requirements.txt
telepot==8.3
Pillow==3.3.0
pyproj
